added sinc histogram, allowing for fractional image source delays

+++ common/include/common/pressure_intensity.h
@@ -16,5 +16,5 @@ inline T intensity_to_pressure(T intensity, U Z) {
     return copysign(sqrt(abs(intensity * Z)), intensity);
 }
 
-float intensity_for_distance(float distance);
-float pressure_for_distance(float distance, float acoustic_impedance);
+double intensity_for_distance(double distance);
+double pressure_for_distance(double distance, double acoustic_impedance);
