+++ common/src/pressure_intensity.cpp
@@ -1,9 +1,9 @@
 #include "common/pressure_intensity.h"
 
-float intensity_for_distance(float distance) {
+double intensity_for_distance(double distance) {
     return 1 / (4 * M_PI * distance * distance);
 }
 
-float pressure_for_distance(float distance, float acoustic_impedance) {
+double pressure_for_distance(double distance, double acoustic_impedance) {
     return std::sqrt(acoustic_impedance / (4 * M_PI)) / distance;
 }
